groq-builder: feature(compat): extracted `projectField` tests to separate file
