+++ packages/groq-builder/src/commands/project.test.ts
@@ -10,116 +10,7 @@ import { currencyFormat } from "../tests/utils";
 import { validate } from "../validation/validate";
 
 const q = createGroqBuilder<SchemaConfig>();
-
 const qVariants = q.star.filterByType("variant");
-
-describe("projectField (naked projections)", () => {
-  const qPrices = qVariants.projectField("price");
-  const qNames = qVariants.projectField("name");
-  const qImages = qVariants.projectField("images[]");
-  const data = mock.generateSeedData({
-    variants: mock.array(5, (i) =>
-      mock.variant({
-        name: `Variant ${i}`,
-        price: 55 + i,
-        msrp: 55 + i,
-      })
-    ),
-  });
-
-  it("can project a number", () => {
-    expectType<InferResultType<typeof qPrices>>().toStrictEqual<
-      Array<number>
-    >();
-    expect(qPrices.query).toMatchInlineSnapshot(
-      '"*[_type == \\"variant\\"].price"'
-    );
-  });
-  it("can project a string", () => {
-    expectType<InferResultType<typeof qNames>>().toStrictEqual<Array<string>>();
-    expect(qNames.query).toMatchInlineSnapshot(
-      '"*[_type == \\"variant\\"].name"'
-    );
-  });
-  it("can project arrays with []", () => {
-    type ResultType = InferResultType<typeof qImages>;
-
-    expectType<ResultType>().toStrictEqual<Array<
-      NonNullable<SanitySchema.Variant["images"]>
-    > | null>();
-  });
-  it("can chain projections", () => {
-    const qSlugCurrent = qVariants.projectField("slug").projectField("current");
-    expectType<InferResultType<typeof qSlugCurrent>>().toStrictEqual<
-      Array<string>
-    >();
-
-    const qImageNames = qVariants
-      .slice(0)
-      .projectField("images[]")
-      .projectField("name");
-    expectType<
-      InferResultType<typeof qImageNames>
-    >().toStrictEqual<Array<string> | null>();
-  });
-
-  it("executes correctly (price)", async () => {
-    const results = await executeBuilder(qPrices, data.datalake);
-    expect(results).toMatchInlineSnapshot(`
-      [
-        55,
-        56,
-        57,
-        58,
-        59,
-      ]
-    `);
-  });
-  it("executes correctly (name)", async () => {
-    const results = await executeBuilder(qNames, data.datalake);
-    expect(results).toMatchInlineSnapshot(`
-      [
-        "Variant 0",
-        "Variant 1",
-        "Variant 2",
-        "Variant 3",
-        "Variant 4",
-      ]
-    `);
-  });
-
-  describe("deep properties", () => {
-    it("invalid entries should have TS errors", () => {
-      // @ts-expect-error ---
-      qVariants.projectField("slug[]");
-      // @ts-expect-error ---
-      qVariants.projectField("slug.INVALID");
-      // @ts-expect-error ---
-      qVariants.projectField("INVALID");
-      // @ts-expect-error ---
-      qVariants.projectField("INVALID.current");
-    });
-
-    it("can project nested properties", () => {
-      const qSlugs = qVariants.projectField("slug.current");
-      expectType<InferResultType<typeof qSlugs>>().toStrictEqual<
-        Array<string>
-      >();
-      expect(qSlugs.query).toMatchInlineSnapshot(
-        '"*[_type == \\"variant\\"].slug.current"'
-      );
-    });
-
-    it("can project arrays with []", () => {
-      const qImages = qVariants.projectField("images[]");
-      type ResultType = InferResultType<typeof qImages>;
-
-      expectType<ResultType>().toStrictEqual<Array<
-        NonNullable<SanitySchema.Variant["images"]>
-      > | null>();
-    });
-  });
-});
 
 describe("project (object projections)", () => {
   const data = mock.generateSeedData({
@@ -661,33 +552,5 @@ describe("project (object projections)", () => {
         })
       );
     });
-  });
-});
-
-describe("grab (backwards compatibility)", () => {
-  it("should be defined", () => {
-    expect(q.grab).toBeTypeOf("function");
-    expect(q.grab$).toBeTypeOf("function");
-    expect(q.grabOne).toBeTypeOf("function");
-    expect(q.grabOne$).toBeTypeOf("function");
-  });
-
-  it("should be type-safe", () => {
-    const qVariants = q.star.filterByType("variant");
-    const qGrab = qVariants.grab((q) => ({
-      name: true,
-      slug: "slug.current",
-      msrp: ["msrp", validate.number()],
-      styles: q.grabOne("style[]").deref().grabOne("name"),
-    }));
-
-    expectType<InferResultType<typeof qGrab>>().toStrictEqual<
-      Array<{
-        name: string;
-        slug: string;
-        msrp: number;
-        styles: Array<string> | null;
-      }>
-    >();
   });
 });
